[TP1-002/3/4] - { feat } :se realizan cambios en el 1103

+++ src/ejercicios/ejercicio1103.py
@@ -23,6 +23,6 @@ if numero1 != 0 and (numero2/numero1)%2 == 0:
     print("Es divisor:",True)
 else:
     print("Es divisor:",False)
-print("El ",numero2,"% de",numero1,"es:",(numero1*numero2/100))
+print("El ",numero2,"% de",numero1,"es:",int(numero1*numero2/100))
 #print("El ", numero2, "% " "de ",numero1, "es", "{:.2f}".format(numero1 * (numero2/100)))
 #print("El precio es $", numero1)
